requirements.txt
RPi.GPIO
requests
git+https://github.com/lthiery/SPI-Py.git
